TSH_Application/backend/requirements.txt
alembic==1.13.1
bcrypt==4.1.2
blinker==1.7.0
boto3==1.28.63
botocore
click==8.1.7
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
python-decouple==3.8
SQLAlchemy==2.0.27
typing_extensions==4.9.0
urllib3
uuid==1.30
Werkzeug==3.0.1
